move out picking ui methods to a separate picking_ui module
